pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
name = "templatr"
dynamic = ["version"]
description = "Templatr library, easy to use templating library."
readme = "README.md"
requires-python = ">=3.8"

authors = [
    { name = "Andres Carrera", email = "andreshcar@live.com" },
]

classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",

    "Operating System :: OS Independent",

    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    
    "License :: OSI Approved :: MIT License",

    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

keywords = ["template"]

dependencies = [
    "pydantic",
    "pyyaml"
]

[project.urls]
Homepage = "https://github.com/carrera-dev-consulting/templatr"
Documentation = "https://consulting.gxldcptrick.dev/templatr/docs"
Repository = "https://github.com/carrera-dev-consulting/templatr"
Issues = "https://github.com/carrera-dev-consulting/templatr/issues"
Changelog = "https://github.com/carrera-dev-consulting/templatr/releases"

[tool.pdm]
version = { source = "file", path = "src/templatr/__init__.py" }
distribution = true


